Modified carts for error messages

+++ src/carts/carts.service.ts
@@ -7,13 +7,27 @@ export class CartsService {
   constructor(private prisma: PrismaService) {}
 
   async addToCart(userId: number, productId: number, quantity: number): Promise<CartItem> {
-    try {
-      const cart = await this.prisma.cart.upsert({
-        where: { userId },
-        update: {},
-        create: { userId },
-      });
 
+    const user = await this.prisma.user.findUnique({ where: { userId } });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    
+    const cart = await this.prisma.cart.upsert({
+      where: { userId },
+      update: {},
+      create: { userId },
+    });
+    if (!cart) {
+      throw new NotFoundException('Cart not found');
+    }
+
+    const product = await this.prisma.product.findUnique({ where: { productId: productId } });
+    if (!product) {
+      throw new NotFoundException('Product not found');
+    }
+
+    try{
       return this.prisma.cartItem.upsert({
         where: {
           cartId_productId: {
@@ -27,19 +41,25 @@ export class CartsService {
     } catch (error) {
       throw new BadRequestException('Failed to add item to cart');
     }
+    
   }
 
-  async removeFromCart(userId: number, productId: number): Promise<void> {
+  async removeFromCart(userId: number, productId: number): Promise<String> {
     const cart = await this.prisma.cart.findUnique({ where: { userId } });
 
     if (!cart) {
       throw new NotFoundException('Cart not found');
     }
 
+    const product = await this.prisma.product.findUnique({ where: { productId } });
+    if (!product) {
+      throw new NotFoundException('Product not found');
+    }
     try {
       await this.prisma.cartItem.delete({
         where: { cartId_productId: { cartId: cart.cartId, productId } },
       });
+      return 'Item removed from cart';
     } catch (error) {
       throw new BadRequestException('Failed to remove item from cart');
     }
@@ -63,6 +83,11 @@ export class CartsService {
 
     if (!cart) {
       throw new NotFoundException('Cart not found');
+    }
+
+    const product = await this.prisma.product.findUnique({ where: { productId } });
+    if (!product) {
+      throw new NotFoundException('Product not found');
     }
 
     try {
